db/requirements.txt
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "3.13"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "3.13"
apscheduler==3.10.4 ; python_version >= "3.10" and python_version < "3.13"
beautifulsoup4==4.12.2 ; python_version >= "3.10" and python_version < "3.13"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "3.13"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
cssselect==1.2.0 ; python_version >= "3.10" and python_version < "3.13"
distro==1.8.0 ; python_version >= "3.10" and python_version < "3.13"
dnspython==2.4.2 ; python_version >= "3.10" and python_version < "3.13"
envparse==0.2.0 ; python_version >= "3.10" and python_version < "3.13"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
feedfinder2==0.0.4 ; python_version >= "3.10" and python_version < "3.13"
feedparser==6.0.11 ; python_version >= "3.10" and python_version < "3.13"
filelock==3.13.1 ; python_version >= "3.10" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.13"
httpcore==1.0.2 ; python_version >= "3.10" and python_version < "3.13"
httpx==0.25.2 ; python_version >= "3.10" and python_version < "3.13"
idna==3.6 ; python_version >= "3.10" and python_version < "3.13"
jieba3k==0.35.1 ; python_version >= "3.10" and python_version < "3.13"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.13"
loguru==0.7.2 ; python_version >= "3.10" and python_version < "3.13"
logutil[loguru]==0.1.5 ; python_version >= "3.10" and python_version < "3.13"
lxml==4.9.3 ; python_version >= "3.10" and python_version < "3.13"
newspaper3k==0.2.8 ; python_version >= "3.10" and python_version < "3.13"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "3.13"
openai==1.5.0 ; python_version >= "3.10" and python_version < "3.13"
pillow==10.1.0 ; python_version >= "3.10" and python_version < "3.13"
pydantic-core==2.14.5 ; python_version >= "3.10" and python_version < "3.13"
pydantic==2.5.2 ; python_version >= "3.10" and python_version < "3.13"
pymongo==4.6.1 ; python_version >= "3.10" and python_version < "3.13"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.13"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
pytz==2023.3.post1 ; python_version >= "3.10" and python_version < "3.13"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
regex==2023.10.3 ; python_version >= "3.10" and python_version < "3.13"
requests-file==1.5.1 ; python_version >= "3.10" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.13"
sgmllib3k==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.13"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
soupsieve==2.5 ; python_version >= "3.10" and python_version < "3.13"
tinysegmenter==0.3 ; python_version >= "3.10" and python_version < "3.13"
tldextract==5.1.1 ; python_version >= "3.10" and python_version < "3.13"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "3.13"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "3.13"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.13" and platform_system == "Windows"
tzlocal==5.2 ; python_version >= "3.10" and python_version < "3.13"
urllib3==2.1.0 ; python_version >= "3.10" and python_version < "3.13"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "3.13" and sys_platform == "win32"
